PI DISP - Add conditional compile elements
