+++ pi-display-foundation/Cargo.toml
@@ -4,7 +4,9 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
+embedded-graphics = "0.8.0"
+
+[target.'cfg(target_os = "linux")'.dependencies]
 sh1106 = "0.5.0"
 linux-embedded-hal = "0.3.0"
-embedded-graphics = "0.8.0"
 tinybmp = "0.6.0"
